# docs/source/openapi.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
  /api/model/:
    get:
      summary: List models
      description: Returns list of models and attributes
      operationId: listModels
      parameters: []
      responses:
        '200':
          description: ''
    post:
      summary: Create model
      description: Add model to the model list
      operationId: createModel
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                system:
                  enum:
                  - optmod
                  - jump
                  - cvxpy
                  - gams
                status:
                  enum:
                  - created
                  - running
                  - success
                  - unknown
                  - error
                source:
                  type: string
                time_created:
                  type: string
                  format: date-time
                time_start:
                  type: string
                  format: date-time
                  nullable: true
                time_end:
                  type: string
                  format: date-time
                  nullable: true
                execution_log:
                  type: string
              required:
              - system
              - time_created
      responses:
        '200':
          description: ''
...
  /api/variable/:
    get:
      operationId: listVariables
      summary: List Variables
      description: Lists all Variables and their properties
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: createVariable
      summary: Create new variable
      description: Adds new variable, returns updated list of variables
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                type:
                  enum:
                  - scalar
                  - array1d
                  - array2d
                  - hashmap
                  - unknown
                size:
                  type: string
                  maxLength: 100
                has_states:
                  type: boolean
              required:
              - name
      responses:
        '200':
          description: ''
...
  /api/function/:
    get:
      operationId: listFunctions
      summary: List all Functions
      description: Returns a list of all Functions
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: createFunction
      summary: Create Function
      description: Creates a new Function, returns updated list
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                type:
                  enum:
                  - scalar
                  - array1d
                  - array2d
                  - hashmap
                  - unknown
                size:
                  type: string
                  maxLength: 100
                has_states:
                  type: boolean
                objective:
                  enum:
                  - scalar
                  - array1d
                  - array2d
                  - hashmap
                  - unknown
              required:
              - name
      responses:
        '200':
          description: ''
...
  /api/constraint/:
    get:
      operationId: listConstraints
      summary: List Constraints
      description: Returns list of all Constraints   
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: createConstraint
      summary: Crate Constraint
      description: Creates new Constraint, returns updated list of all Constraints
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                type:
                  enum:
                  - scalar
                  - array1d
                  - array2d
                  - array3d
                  - hashmap
                  - unknown
                size:
                  type: string
                  maxLength: 100
                has_states:
                  type: boolean
              required:
              - name
      responses:
        '200':
          description: ''
...
  /api/problem/:
    get:
      operationId: listProblems
      summary: List Problems
      description: Returns a list of all problems
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: createProblem
      summary: Create Problem
      description: Creates new problem, returns updatd list of all Problems
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 100
              required:
              - name
      responses:
        '200':
          description: ''
...
  /api/solver/:
    get:
      operationId: listSolvers
      summary: List Solvers
      description: Returns a list of all Solvers
      parameters: []
      responses:
        '200':
          description: ''
    post:
      operationId: createSolver
      summary: Create Solver
      description: Creates new Solver, returns updated list of Solvers      
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: ''
...